double, bez ostatniego, dalej nie

+++ include/fun.h
@@ -33,6 +33,6 @@ wektor iloczynwektorowy(wektor w1, wektor w2);
 double iloczynskalarny(wektor w1, wektor w2);
 bool wektoryxx(wektor w1, wektor w2);
 bool wektoryxxxx(wektor w1, wektor w2);
-int losowa_ab(int a, int b);
+double losowa_ab(double a, double b);
 
 #endif
